requirements.txt.txt
asgiref==3.2.10
click==7.1.2
cmake==3.17.3
Django==3.0.14
django-filter==2.3.0
dlib==19.8.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1